fix: close eleven trivial bugs across formats, filesystem and cli
- documentation attributes move from core into the documentation tool
- mago guard stops lib packages from importing core
- one readLines() contract for every stream, S3 and Azure included
- nested parquet decimals are writable and sized by the spec
- CLI commands name themselves with #[AsCommand]
- blank JSON lines, null DBAL platform options and empty files no longer break reads or schema inference

+++ src/core/etl/src/Flow/ETL/Function/Average.php
@@ -122,7 +122,7 @@ final class Average implements AggregatingFunction, FrameAccumulating, WindowFun
             return null;
         }
 
-        return (float) (new Calculator())->divide($this->sum, $this->count, $this->scale, $this->rounding);
+        return (new Calculator())->divide($this->sum, $this->count, $this->scale, $this->rounding);
     }
 
     public function toString(): string
